remove references
