+++ crates/ra_hir/src/code_model_impl.rs
@@ -1,3 +1,2 @@
 mod konst; // `const` is invalid ident :(
-mod module;
 pub(crate) mod function;
